Rearrange some pytet fixtures.
