+++ tests/test_rp_exec.py
@@ -69,23 +69,6 @@ def test_rp_usability():
             ...
 
 
-@pytest.fixture(scope='session')
-def rpsession():
-    import radical.pilot as rp
-    # Note: Session creation will fail with a FileNotFound error unless venv
-    #       is explicitly `activate`d (or the scripts installed with RADICAL components
-    #       are otherwise made available on the PATH).
-
-    # Note: radical.pilot.Session creation causes several deprecation warnings.
-    # Ref https://github.com/radical-cybertools/radical.pilot/issues/2185
-    with warnings.catch_warnings():
-        warnings.simplefilter('ignore', category=DeprecationWarning)
-        session = rp.Session()
-    with session:
-        yield session
-    assert session.closed
-
-
 @with_radical_only
 def test_rp_basic_task(rpsession):
